requirements.txt
# dev requirements, prod requirements are found in setup.py
#

frigg-common
redis
raven
docker-py

# testing tools
pytest
pytest-cov
mock
responses
